Add build task to wrap CommonJS modules in AMD
As part of grunt deps, wraps NodeJS-style CommonJS modules in AMD for
usage with require.js.

To wrap a CommonJS module, include the module's path in the amdWrap
property of require.config.js. The destination for wrapped modules is
libs/amd-wrap. Use the wrapped module in the paths section of
require.config.js. This step is only necessary for modules on npm that
do not publish an AMD or a UMD style module i.e. the module assumes
either exports or module.exports is available on global scope.