Dashboard: Fix date histogram errors
Simplifies date histogram aggregation initial interval calculation which
was causing an error to render instead of the dashboard card.

Includes the interval unit (ms) sent to the backend to prevent
ElasticSearch 1.x returning an error when the interval is a float and to
prepare for Elastic 2.x, which requires all durations to specify the
unit of time.

v5analytics/visallo-lts#793