Fix flashing of right aligned numbers in multiple selection histogram detail pane.
The svg is set to be 100% of the detail pane width, IE doesn't update those elements sizes until after mouseup. So the numbers were being pushed offscreen for a little.

Use a media query to target ie 10+ to fix IE11, and force the scrollbar to stick around so the width isn't changing based on scrollbar visibility.