Changed: ACLProvider: support elements which cannot be found
If an element cannot be found for what ever reason return to the client
a readonly ClientApiElement. Element not found could be caused by a blind
write edge which does not yet have it's other side yet, without this
an exception is thrown and the UI becomes unusable.

CHANGELOG
Changed: ACLProvider: support elements which cannot be found