Added: WorkerBase: add support to not terminate if nextTuple fails
In some scenarios such as running the workers from your web servers
you may not want a worker to stop completely on exception while
retrieving the next tuple from the spout. This commit adds support to
continue on failure.

CHANGELOG
Added: WorkerBase: add support to not terminate if nextTuple fails