Changed: Move visibility in the details pane into the info box
This commit moves the visibility in the details pane for both
properties and comments from below the value to the popup info
box.

When using long visibility strings the details pane can become
cluttered quite fast. This addresses this by moving what most
users will not not typically use to get the job down into the
info box which is just one click away.

CHANGELOG
Changed: Move visibility in the details pane into the info box