Added: Search realted from related entities in detail pane
If the related items in the details pane is very large it
might be nice to search those related entities and filter them
down. This commit adds a link to the related entities in the
details pane to open in search.

CHANGELOG
Added: Search realted from related entities in detail pane