Update the script to install the archetype jar
The archetype jar needs to be made before running. Instead of having a
person who is working on the project run it, I instead just generate it
in the script before creating the project. Less steps for everyone and
it only adds a second or two to the plugin generate time.