Find correct gpw property to process
Whenever the value of a published property has been changed on a
workspace, those changes were not being put through the graph property
workers. This change checks all of the properties that can be returned
and uses the workspace ID to decide which property needs to be put
through the graph property workers.

format code

fix tests

changes based on joes suggestion