Fixed: work product ontology speedup when running unit tests (#906)
Previously a static field was used to track the ontology updates, this
caused problems when running unit tests which would access the classes
multiple times clearing the ontology in between. This commit changes
this behavior to look up the ontology items in the ontology repository
which has the correct state.