Web: Snap to grid graph position fixes
Fixes issues dropping an element onto the graph when snap to grid is
enabled and elements snapping back to the current position if not moved
to a new position on the grid.

Moves position change logic when snap to grid is on out of the reducer
and into the action creator function to prevent side effects.