Store: Add support for undo/redo redux async actions
Stores actions that can be undone or redone under the undoActionHistory
location in redux in undos and redos arrays. To mark an action as
undoable and redoable, include undo and redo properties in the action's
payload along with an optional scope, which defaults to global. When an
UNDO or REDO action is fired, the redux undo middleware checks the undos
or redos array respectively that matches the action's scope and
dispatches the appropriate action creator function registered with the
plugin repository. Plugins may register undo and redo handlers by
registering an org.visallo.store extension:

registry.registerExtension('org.visallo.store', {
    undoActions: {
      [ACTION_TYPE]: {
          undo: (undo) => {
              //handle undo
          },
          redo: (redo) => {
              //handle redo
          }
      }
    }
});

The above approach is designed to support handling undo actions that
trigger async operations with the Visallo server and the potentially
large size of Visallo's state tree. Typical redux undo approaches are
designed to undo handle synchronous undos and snapshot the entire state
tree on every change.

Removes the no longer used UndoManager.