Changed: StyleAppendableHandler to compile less on append
StyleAppendableHandler is only used when devMode is false so we don't need
to support reloading of less and css. This change compiles and appends the less
css to a string. This should decrease the time needed to serve up primary css of
the application since we don't have to compile the less each time.