Changed: /vertex/edges to use Vertexium vertex query for performance (#1151)
Before this commit the vertex was retrieved and all edges were evaluated to
determine if they should be included in the response. This change switches to
using a vertex query which will use ElasticSearch under the covers to quickly
find and page results.

This change will also facilitate sorting by some edge property in the future
instead of key order.