Ignore elements with 'ui-ignore-pane-width' class when calculating available width.
Panes could sometimes be set to smaller  widths because when calculating the available width we were including descendant resizable elements in the calculation (e.g. the multi-detail panel was counted twice once for the pane's width and once for the resizable element list's width even though the list was inside the pane).